requirements.txt
absl-py==2.3.1
astunparse==1.6.3
boto3==1.42.3
botocore==1.42.3
cachetools==6.2.2
certifi==2025.11.12
charset-normalizer==3.4.4
cmdstanpy==1.3.0
contourpy==1.3.0
cycler==0.12.1
flatbuffers==25.9.23
fonttools==4.60.1
gast==0.4.0
google-auth==2.43.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.76.0
h5py==3.14.0
holidays==0.83
idna==3.11
importlib_metadata==8.7.0
importlib_resources==6.5.2
jax==0.4.30
jaxlib==0.4.30
jmespath==1.0.1
joblib==1.5.2
keras==2.12.0
kiwisolver==1.4.7
libclang==18.1.1
Markdown==3.9
MarkupSafe==3.0.3
matplotlib==3.9.4
ml_dtypes==0.5.4
oauthlib==3.3.1
opt_einsum==3.4.0
packaging==25.0
pandas==2.3.3
pillow==11.3.0
prophet==1.2.1
protobuf==4.25.8
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
s3transfer==0.16.0
scikit-learn==1.6.1
scipy==1.13.1
six==1.17.0
sklearn==0.0
stanio==0.5.1
tensorboard==2.12.3
tensorboard-data-server==0.7.2
tensorflow-cpu @ file:///home/ec2-user/tensorflow_cpu-2.12.0-cp39-cp39-manylinux_2_17_x86_64.manylinux2014_x86_64.whl
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==3.1.0
threadpoolctl==3.6.0
tqdm==4.67.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==1.26.20
Werkzeug==3.1.4
wrapt==1.14.2
zipp==3.23.0
